Got the frontend working again
