--- Directory.Build.targets
@@ -1,13 +1,0 @@
-<!-- 
-  This file is imported late in the build order. 
-  Use it to override properties and define dependent properties.
--->
-<Project>
-  <PropertyGroup>
-    <MSBuildProjectDirRelativePath>$([MSBuild]::MakeRelative($(RepoRoot), $(MSBuildProjectDirectory)))</MSBuildProjectDirRelativePath>
-    <NodeModulesRelativePath>$([MSBuild]::MakeRelative($(MSBuildProjectDirectory), $(RepoRoot)))</NodeModulesRelativePath>
-  </PropertyGroup>
-  <Target Name="CheckNxModuleBoundaries" BeforeTargets="Build">
-    <Exec Command="node $(NodeModulesRelativePath)/node_modules/@nx-dotnet/core/src/tasks/check-module-boundaries.js --project-root &quot;$(MSBuildProjectDirRelativePath)&quot;"/>
-  </Target>
-</Project>
