--- apps/frontend/project.json
@@ -1,8 +1,0 @@
-{
-  "name": "frontend",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "apps/frontend",
-  "projectType": "application",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project frontend --web"
-}
