+++ apps/frontend/src/_services/buildQuizData.ts
@@ -1,7 +1,7 @@
-import quizPaths from '_data/constants/paths.json';
-import quizAnswers from '_data/constants/answers.json';
-import quizTexts from '_data/constants/extendedTexts.json';
-import { QuizPath, QuizAnswer, QuizText, Node } from '_data/types/types';
+import quizPaths from '@/_data/constants/paths.json';
+import quizAnswers from '@/_data/constants/answers.json';
+import quizTexts from '@/_data/constants/extendedTexts.json';
+import { QuizPath, QuizAnswer, QuizText, Node } from '@/_data/types/types';
 
 export function buildQuizData(): Node {
   const paths: QuizPath = quizPaths as QuizPath;
