+++ apps/frontend/src/app/quiz/page.tsx
@@ -1,5 +1,5 @@
-import QuizPage from '_components/Quiz/quiz';
-import { getHistoryFunction } from '_utils/quiz-functions';
+import QuizPage from '@/_components/Quiz/quiz';
+import { getHistoryFunction } from '@/_utils/quiz-functions';
 import { getSession } from '@auth0/nextjs-auth0';
 
 export default async function Page() {
