--- package.json
@@ -1,69 +1,0 @@
-{
-  "name": "@finbud/source",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {},
-  "private": true,
-  "dependencies": {
-    "@auth0/nextjs-auth0": "^3.6.0",
-    "@tabler/icons-react": "^3.31.0",
-    "bootstrap": "^5.3.3",
-    "next": "^15.2.3",
-    "react": "19.0.0",
-    "react-bootstrap": "^2.10.9",
-    "react-dom": "19.0.0",
-    "react-icons": "^5.5.0",
-    "react-spinners": "^0.15.0",
-    "react-type-animation": "^3.2.0",
-    "tailwind-merge": "^3.0.2"
-  },
-  "devDependencies": {
-    "@eslint/compat": "^1.1.1",
-    "@eslint/eslintrc": "^2.1.1",
-    "@eslint/js": "^9.8.0",
-    "@nx-dotnet/core": "^2.5.0",
-    "@nx/devkit": "20.6.2",
-    "@nx/eslint": "20.6.2",
-    "@nx/eslint-plugin": "20.6.2",
-    "@nx/jest": "20.6.2",
-    "@nx/js": "20.6.2",
-    "@nx/next": "20.6.2",
-    "@nx/playwright": "20.6.2",
-    "@nx/workspace": "20.6.2",
-    "@playwright/test": "^1.36.0",
-    "@swc-node/register": "~1.9.1",
-    "@swc/core": "~1.5.7",
-    "@swc/helpers": "~0.5.11",
-    "@testing-library/dom": "10.4.0",
-    "@testing-library/react": "16.1.0",
-    "@types/jest": "^29.5.12",
-    "@types/node": "18.16.9",
-    "@types/react": "19.0.0",
-    "@types/react-dom": "19.0.0",
-    "autoprefixer": "10.4.13",
-    "babel-jest": "^29.7.0",
-    "daisyui": "^4.12.10",
-    "eslint": "^9.8.0",
-    "eslint-config-next": "^15.2.3",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "2.31.0",
-    "eslint-plugin-jsx-a11y": "6.10.1",
-    "eslint-plugin-playwright": "^1.6.2",
-    "eslint-plugin-react": "7.35.0",
-    "eslint-plugin-react-hooks": "5.0.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "nx": "20.6.2",
-    "postcss": "8.4.38",
-    "prettier": "^2.6.2",
-    "tailwindcss": "3.4.3",
-    "ts-jest": "^29.1.0",
-    "ts-node": "10.9.1",
-    "tslib": "^2.3.0",
-    "typescript": "~5.7.2",
-    "typescript-eslint": "^8.19.0"
-  },
-  "workspaces": [
-    "apps/*"
-  ]
-}
